+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # Changelog
 
+## 0.3.0 - 2017-01-21
+
+- Add coordinate precision as an option
+- Add dropZ option
+
 ## 0.2.0 - 2016-11-25
 
 - Remove empty properties when saving GeoJSON